requirements.txt
Django>=3.2.10,<3.3.0
djangorestframework>=3.13.1,<3.14.0